refactor workbench risk routes
